+++ CHANGELOG.md
@@ -6,6 +6,7 @@
      bridges.
    - Add "as" field as a copy of the "as_number" field in preparation
      of removing the "as_number" field in the future.
+   - Add new "as_name" parameter to search relays by AS name.
 
 
 # Changes in version 6.1-1.15.0 - 2018-07-16